refs #5711 - [Widget-Native-Android] Make a possibility to control widget width/height from RMS

+++ widget-app/src/main/java/com/github/funler/widget_android/WidgetViewActivity.java
@@ -70,7 +70,7 @@ public class WidgetViewActivity extends AppCompatActivity {
             double webViewHeight = dpFromPx(getBaseContext(), bridgeWebView.getHeight());
             double margin = (windowHeight - webViewHeight);
 
-            double y = intent.getExtras().getDouble("y") + margin;
+            double y = intent.getExtras().getFloat("y") + margin;
 
             double visibleHeight = windowHeight / 2;
